pyproject.toml
[project]
name = "flappy-rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "neat-python>=2.0.0",
    "pydantic-settings>=2.13.1",
    "pygame-ce>=2.5.7",
    "tomli>=2.4.0",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]
